requirements.txt
discord.py
requests
asyncio
numpy
PyNaCl
youtube-dl
ffmpeg